Makefile: Silence some overly verbose warnings by GCC
Compiler warnings like address, array-compare, format, and
implicit-fallthrough are the results of GCC's notorious aggressiveness.
They are harmless and can be ignored. GCC can take care of them and
optimize them accordingly. So, silence them to get rid of the insane
verbosity.

implicit-fallthrough is disabled for clang by default. So, we can safely
remove the explicit cc-option for implicit-fallthrough as well.

Change-Id: b121e89b4769ad4016d067e29dedee55bea88b06
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>